Massively refactor entire source base
Proof of concepts were removed from tracking. Various proofing code
  throughout important files (a la WordsPreProcessor.kt) was removed

A lot of code was just "Java in Kotlin syntax", which is pretty poor: we
  can do better and have done so. Things like the usage of 'let' in the
  ThreadedWorker class.

Accidentally (and helpfully) formatted the JSON files, so they can be
  investigated with the human eye now ;)

There is a message queue system within "LocalBulletinBoard", but it
  isn't utilized anywhere.

Removed all usages of the "four letter word lists", save in the use in
  WordGenerator.
HashingUtilities inherited the function from a POC.
Functions were renamed. Instance constants were made class constants
  (via companion objects)